move API definitions to separate module
